footium_mcp_server/requirements.txt
# Core dependencies
httpx>=0.24.0
mcp>=0.1.0
gql>=3.4.0
graphql-core>=3.2.0

# Development dependencies
black>=23.1.0
pylint>=2.16.0
